refactor: Dao 중복 코드를 제거하기 위해 상위 추상 클래스인 BaseDao 생성
